Added ready system
Added a ready system with some first changes. Unfortunately I could not test the ready system as I bound it on shooting and I cannot shoot with the 4th character on my laptop :(. Disabled the weapon system for now in endgame. These two could work together as when someone shoots, the characters readys up too.